add decimal

+++ bond_service/bond_service_api/utils.py
@@ -3,16 +3,18 @@ from django.db.models.query import QuerySet
 from typing import Dict
 from django.utils import timezone
 from .serializers import BondSerializer
+from decimal import Decimal
 
 
 def get_portfolio_analysis(portfolio: Portfolio) -> Dict:
     all_portfolio_bonds: QuerySet[Bond] = portfolio.bonds.all()
     if all_portfolio_bonds.exists():
-        total_value = sum(all_portfolio_bonds.values_list('bond_value', flat=True))
-        avg_interest_rate = sum(all_portfolio_bonds.values_list(
-            'interest_rate', flat=True)) / all_portfolio_bonds.count()
+        total_value = Decimal(sum(all_portfolio_bonds.values_list('bond_value', flat=True)))
+        avg_interest_rate = Decimal(sum(all_portfolio_bonds.values_list(
+            'interest_rate', flat=True)) / all_portfolio_bonds.count())
         nearest_maturity_bond: Bond = min(all_portfolio_bonds, key=lambda bond: bond.maturity_date)
-        years = (nearest_maturity_bond.maturity_date - timezone.now().date()).days / 365.25
+        years = Decimal((nearest_maturity_bond.maturity_date - timezone.now().date()).days / Decimal(365.25))
+
         future_value = total_value * (1 + avg_interest_rate / 100) ** years
 
         return {
